Refining Tests
I added a dependency on XogenyTest in order to run the test cases.  I should probably include some examples or something that aren't strictly tests so at least people can see how the models work, even if they don't have the testing library.

I tried my best to deal with initial condition warnings, but this is tricky in the case of the kinematic clutch models, so some still exist.

I refined the KinematicClutch model a bit to handle initialization better by putting the intuitive initial conditions in depending on the value of the command flag.