Fixed an issue raised by @msasena
@msasena pointed out that checking this library with Dymola 2013 (with
simulation) gave an error.  While the translational error is intentional
and was meant to demonstrate an issue with a naive model, I shouldn't
have included an experiment annotation since this gives Dymola the
impression that the model is expected to simulate.

I've removed the experiment annotation and added some documentation to
the model that cannot simulate explaining that the model is not intended
to translate and provided a link to the blog post where this is
discussed.